Add gameplay functions
makeMove will first check to see if the card has
already been selected, if so, this will lead to
game over. If it's a valid move, a point is
added to the score and the selected card is
placed in the the selectedCards array.

checkForLoss returns true or false if the card
is in the selectedCards array.